+++ projects/start-os/web/ui/src/app/services/api/embassy-live-api.service.ts
@@ -31,9 +31,11 @@ import {
   PkgAddPrivateDomainReq,
   PkgAddPublicDomainReq,
   PkgBindingSetAddressEnabledReq,
+  PkgBindingSetGuaAccessReq,
   PkgRemovePrivateDomainReq,
   PkgRemovePublicDomainReq,
   ServerBindingSetAddressEnabledReq,
+  ServerBindingSetGuaAccessReq,
   ServerState,
   WebsocketConfig,
 } from './api.types'
@@ -678,6 +680,24 @@ export class LiveApiService extends ApiService {
   ): Promise<null> {
     return this.rpcRequest({
       method: 'package.host.binding.set-range-address-enabled',
+      params,
+    })
+  }
+
+  async serverBindingSetGuaAccess(
+    params: ServerBindingSetGuaAccessReq,
+  ): Promise<null> {
+    return this.rpcRequest({
+      method: 'server.host.binding.set-gua-access',
+      params,
+    })
+  }
+
+  async pkgBindingSetGuaAccess(
+    params: PkgBindingSetGuaAccessReq,
+  ): Promise<null> {
+    return this.rpcRequest({
+      method: 'package.host.binding.set-gua-access',
       params,
     })
   }